requirements.txt
fastapi==0.89.1
uvicorn==0.20.0
SQLAlchemy==2.0.0
ipdb==0.13.11
pytest==7.2.1
httpx==0.23.3
PyJWT==2.6.0